# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LearningResponse {
  id: number;
  /** The user's response text, or empty string if "I don't know" */
  text: string;
  /** True if the user tapped "I don't know" */
  skipped: boolean;
  createdAt: string;
  /** AI-generated quiz question for this entry, or empty string if not yet generated */
  question?: string;
}

export interface CreateResponseRequest {
  text: string;
  skipped: boolean;
}

export interface UpdateResponseRequest {
  text: string;
}

export type ResponseStatsDaysItem = {
  /** ISO date (YYYY-MM-DD) */
  date: string;
  count: number;
};

export interface ResponseStats {
  total: number;
  answered: number;
  skipped: number;
  days: ResponseStatsDaysItem[];
}

export interface Settings {
  /**
   * @minimum 1
   * @maximum 240
   */
  reminderIntervalMinutes: number;
  quietHoursEnabled: boolean;
  /**
   * Start of quiet window in HH:mm (24h)
   * @pattern ^([01]\d|2[0-3]):[0-5]\d$
   */
  quietHoursStart: string;
  /**
   * End of quiet window in HH:mm (24h)
   * @pattern ^([01]\d|2[0-3]):[0-5]\d$
   */
  quietHoursEnd: string;
}

export type CardReviewRequestResult =
  (typeof CardReviewRequestResult)[keyof typeof CardReviewRequestResult];

export const CardReviewRequestResult = {
  remembered: "remembered",
  forgot: "forgot",
} as const;

export interface CardReviewRequest {
  result: CardReviewRequestResult;
}

export interface UpdateSettingsRequest {
  /**
   * @minimum 1
   * @maximum 240
   */
  reminderIntervalMinutes?: number;
  quietHoursEnabled?: boolean;
  /** @pattern ^([01]\d|2[0-3]):[0-5]\d$ */
  quietHoursStart?: string;
  /** @pattern ^([01]\d|2[0-3]):[0-5]\d$ */
  quietHoursEnd?: string;
}

export interface ErrorResponse {
  message: string;
}