# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /responses:
    get:
      operationId: listResponses
      tags: [responses]
      summary: List all learning responses
      description: Returns all responses, newest first
      responses:
        "200":
          description: List of responses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LearningResponse"
    post:
      operationId: createResponse
      tags: [responses]
      summary: Create a new learning response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResponseRequest"
      responses:
        "201":
          description: Created response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningResponse"
  /responses/{responseId}:
    patch: 
      operationId: updateResponse
      tags: [responses]
      summary: Updates the existing journal entry
      parameters:
        - name: responseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content: 
          application/json:
            schema:
               $ref: "#/components/schemas/UpdateResponseRequest"
      responses:
        "200":
          description: Updated response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Response not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    LearningResponse:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
          description: The user's response text, or empty string if "I don't know"
        skipped:
          type: boolean
          description: True if the user tapped "I don't know"
        createdAt:
          type: string
          format: date-time
        question:
          type: string
          description: AI-generated quiz question for this entry, or empty string if not yet generated
      required:
        - id
        - text
        - skipped
        - createdAt
    CreateResponseRequest:
      type: object
      properties:
        text:
          type: string
        skipped:
          type: boolean
      required:
        - text
        - skipped
    UpdateResponseRequest:
      type: object
      properties:
        text:
          type: string
      required:
        - text
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message